ISS51 - add debug mode
