+++ python_src/models/simulation_data/__init__.py
@@ -1,3 +1,4 @@
 from .vehicle_state_reader import VehicleStateReader
 from .simulation_state_reader import SimulationDataCollector
 from .simulation_data import SimulationParams, SimulationDataRecords, SimulationData, SimulationDataRecord
+from .simulation_data_container import SimulationDataContainer
